add user info data to table and changed data types
Can now add user info data to the USER_INFO table in the database via the pop up window widget. Also made changes where i forgot to add the user_id value to the table and also made it a primary key. Made changes so that the pay value data type is now an integer type since it's the credit card digit integer