added rest of tables, only theater has gui
rest of tables are added in the db with pks, fks, and not null attributes. Only theater has gui window for now. Select which table to pop up gui when running db_connect file. Enter 1 or 2 for now in output terminal. OR simply choose which table file to run. Could not add history or ticket entries : ERRORS!!